+++ ui/client/http/HttpService.ts
@@ -230,6 +230,11 @@ export default {
       })
   },
 
+  validateNode(processId, node) {
+    return api.post(`/nodes/${processId}/validation`, node)
+      .catch(error => this.addError("Failed to get node validation", error, true))
+  },
+
   getNodeAdditionalData(processId, node) {
     return api.post(`/nodes/${processId}/additionalData`, node)
       .catch(error => this.addError("Failed to get node additional data", error, true))