+++ ui/restmodel/src/main/scala/pl/touk/nussknacker/restmodel/displayedgraph/DisplayableProcess.scala
@@ -1,10 +1,10 @@
 package pl.touk.nussknacker.restmodel.displayedgraph
 
 import io.circe.generic.JsonCodec
+import pl.touk.nussknacker.engine.ProcessingTypeData.ProcessingType
 import pl.touk.nussknacker.engine.api.{MetaData, ProcessAdditionalFields, TypeSpecificData}
 import pl.touk.nussknacker.engine.graph.exceptionhandler.ExceptionHandlerRef
 import pl.touk.nussknacker.engine.graph.node.NodeData
-import pl.touk.nussknacker.engine.ProcessingTypeData.ProcessingType
 import pl.touk.nussknacker.restmodel.displayedgraph.displayablenode._
 import pl.touk.nussknacker.restmodel.validation.ValidationResults.ValidationResult
 import pl.touk.nussknacker.engine.graph.NodeDataCodec._
@@ -13,31 +13,25 @@ import pl.touk.nussknacker.engine.graph.NodeDataCodec._
 //or we'd have to do composition which would break many things in client
 // todo: id type should be ProcessName
 @JsonCodec case class DisplayableProcess(id: String,
-                              properties: ProcessProperties,
-                              nodes: List[NodeData],
-                              edges: List[Edge],
-                              processingType: ProcessingType) {
+                                         properties: ProcessProperties,
+                                         nodes: List[NodeData],
+                                         edges: List[Edge],
+                                         processingType: ProcessingType) {
+
+  val metaData: MetaData = properties.toMetaData(id)
 
   def withSuccessValidation(): ValidatedDisplayableProcess = {
     new ValidatedDisplayableProcess(this, ValidationResult.success)
   }
 
-  val metaData = MetaData(
-    id = id,
-    typeSpecificData = properties.typeSpecificProperties,
-    isSubprocess = properties.isSubprocess,
-    additionalFields = properties.additionalFields,
-    subprocessVersions = properties.subprocessVersions
-  )
-
 }
 
 @JsonCodec case class ValidatedDisplayableProcess(id: String,
-                                       properties: ProcessProperties,
-                                       nodes: List[NodeData],
-                                       edges: List[Edge],
-                                       processingType: ProcessingType,
-                                       validationResult: ValidationResult) {
+                                                  properties: ProcessProperties,
+                                                  nodes: List[NodeData],
+                                                  edges: List[Edge],
+                                                  processingType: ProcessingType,
+                                                  validationResult: ValidationResult) {
 
   def this(displayableProcess: DisplayableProcess, validationResult: ValidationResult) =
     this(
@@ -49,7 +43,7 @@ import pl.touk.nussknacker.engine.graph.NodeDataCodec._
       validationResult
     )
 
-  def toDisplayable = DisplayableProcess(id, properties, nodes, edges, processingType)
+  def toDisplayable: DisplayableProcess = DisplayableProcess(id, properties, nodes, edges, processingType)
 
 }
 
@@ -57,5 +51,13 @@ import pl.touk.nussknacker.engine.graph.NodeDataCodec._
                                         exceptionHandler: ExceptionHandlerRef,
                                         isSubprocess: Boolean = false,
                                         additionalFields: Option[ProcessAdditionalFields] = None,
-                                        subprocessVersions: Map[String, Long]
-                            )
+                                        subprocessVersions: Map[String, Long] = Map.empty) {
+
+  def toMetaData(id: String): MetaData = MetaData(
+    id = id,
+    typeSpecificData = typeSpecificProperties,
+    isSubprocess = isSubprocess,
+    additionalFields = additionalFields,
+    subprocessVersions = subprocessVersions
+  )
+}