+++ docs/Changelog.md
@@ -20,6 +20,7 @@ Nussknacker versions
 * [#976](https://github.com/TouK/nussknacker/pull/976) Fixed save button & groups expand for businessView
 * [#973](https://github.com/TouK/nussknacker/pull/973) Textarea editor
 * Introduction to KafkaAvro API: [#871](https://github.com/TouK/nussknacker/pull/871), [#881](https://github.com/TouK/nussknacker/pull/881), [#903](https://github.com/TouK/nussknacker/pull/903), [#981](https://github.com/TouK/nussknacker/pull/981)
+* [#978](https://github.com/TouK/nussknacker/pull/978) Dynamic filter validation
 
 0.1.2
 ------------------------